refactor: fixed layout of project
